// packages/blockchain/typechain-types/EscrowEngine.ts
export declare namespace EscrowEngine {
  export type EscrowStruct = {
    agreementId: string;
    subscriber: AddressLike;
    provider: AddressLike;
    amount: BigNumberish;
    status: BigNumberish;
    proofHash: string;
  };

  export type EscrowStructOutput = [
    agreementId: string,
    subscriber: string,
    provider: string,
    amount: bigint,
    status: bigint,
    proofHash: string
  ] & {
    agreementId: string;
    subscriber: string;
    provider: string;
    amount: bigint;
    status: bigint;
    proofHash: string;
  };
}

export interface EscrowEngineInterface extends Interface {
  getFunction(
    nameOrSignature: "escrows" | "getEscrow" | "lockFunds" | "releaseFunds"
  ): FunctionFragment;

  getEvent(
    nameOrSignatureOrTopic: "FundsLocked" | "FundsReleased"
  ): EventFragment;

  encodeFunctionData(functionFragment: "escrows", values: [string]): string;
  encodeFunctionData(functionFragment: "getEscrow", values: [string]): string;
  encodeFunctionData(
    functionFragment: "lockFunds",
    values: [string, AddressLike]
  ): string;
  encodeFunctionData(
    functionFragment: "releaseFunds",
    values: [string, string]
  ): string;

  decodeFunctionResult(functionFragment: "escrows", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getEscrow", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "lockFunds", data: BytesLike): Result;
  decodeFunctionResult(
    functionFragment: "releaseFunds",
    data: BytesLike
  ): Result;
}

export namespace FundsLockedEvent {
  export type InputTuple = [
    agreementId: string,
    subscriber: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [
    agreementId: string,
    subscriber: string,
    amount: bigint
  ];
  export interface OutputObject {
    agreementId: string;
    subscriber: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export namespace FundsReleasedEvent {
  export type InputTuple = [
    agreementId: string,
    provider: AddressLike,
    amount: BigNumberish
  ];
  export type OutputTuple = [
    agreementId: string,
    provider: string,
    amount: bigint
  ];
  export interface OutputObject {
    agreementId: string;
    provider: string;
    amount: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}

export interface EscrowEngine extends BaseContract {
  connect(runner?: ContractRunner | null): EscrowEngine;
  waitForDeployment(): Promise<this>;

  interface: EscrowEngineInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  escrows: TypedContractMethod<
    [arg0: string],
    [
      [string, string, string, bigint, bigint, string] & {
        agreementId: string;
        subscriber: string;
        provider: string;
        amount: bigint;
        status: bigint;
        proofHash: string;
      }
    ],
    "view"
  >;

  getEscrow: TypedContractMethod<
    [_agreementId: string],
    [EscrowEngine.EscrowStructOutput],
    "view"
  >;

  lockFunds: TypedContractMethod<
    [_agreementId: string, _provider: AddressLike],
    [void],
    "payable"
  >;

  releaseFunds: TypedContractMethod<
    [_agreementId: string, _proofHash: string],
    [void],
    "nonpayable"
  >;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "escrows"
  ): TypedContractMethod<
    [arg0: string],
    [
      [string, string, string, bigint, bigint, string] & {
        agreementId: string;
        subscriber: string;
        provider: string;
        amount: bigint;
        status: bigint;
        proofHash: string;
      }
    ],
    "view"
  >;
  getFunction(
    nameOrSignature: "getEscrow"
  ): TypedContractMethod<
    [_agreementId: string],
    [EscrowEngine.EscrowStructOutput],
    "view"
  >;
  getFunction(
    nameOrSignature: "lockFunds"
  ): TypedContractMethod<
    [_agreementId: string, _provider: AddressLike],
    [void],
    "payable"
  >;
  getFunction(
    nameOrSignature: "releaseFunds"
  ): TypedContractMethod<
    [_agreementId: string, _proofHash: string],
    [void],
    "nonpayable"
  >;

  getEvent(
    key: "FundsLocked"
  ): TypedContractEvent<
    FundsLockedEvent.InputTuple,
    FundsLockedEvent.OutputTuple,
    FundsLockedEvent.OutputObject
  >;
  getEvent(
    key: "FundsReleased"
  ): TypedContractEvent<
    FundsReleasedEvent.InputTuple,
    FundsReleasedEvent.OutputTuple,
    FundsReleasedEvent.OutputObject
  >;

  filters: {
    "FundsLocked(string,address,uint256)": TypedContractEvent<
      FundsLockedEvent.InputTuple,
      FundsLockedEvent.OutputTuple,
      FundsLockedEvent.OutputObject
    >;
    FundsLocked: TypedContractEvent<
      FundsLockedEvent.InputTuple,
      FundsLockedEvent.OutputTuple,
      FundsLockedEvent.OutputObject
    >;

    "FundsReleased(string,address,uint256)": TypedContractEvent<
      FundsReleasedEvent.InputTuple,
      FundsReleasedEvent.OutputTuple,
      FundsReleasedEvent.OutputObject
    >;
    FundsReleased: TypedContractEvent<
      FundsReleasedEvent.InputTuple,
      FundsReleasedEvent.OutputTuple,
      FundsReleasedEvent.OutputObject
    >;
  };
}